requirements.txt
Django==2.2
python-decouple==3.3
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
gunicorn==20.0.4
dj-database-url==0.5.0
psycopg2-binary==2.8.5
